requirements.txt
Django==3.1
djangorestframework==3.11.1
joblib==0.16.0
numpy==1.19.1
pandas==1.1.0
scikit-learn==0.23.1
jupyter==1.0.0